Migration/users

+++ database/migrations/2024_02_21_140410_create_time_and_material_contracts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('contract_id')->constrained('contracts');
             $table->text('milestone_desc');
             $table->date('milestone_enddate');
-            $table->decimal('amount');
+            $table->double('amount');
             $table->timestamps();
         });
     }
